requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
chromadb
faiss-cpu
cassio
tiktoken
datasets